feat: Vanilla JavaScript로 SPA 구현하기
