+++ ToDoMate-SPA/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "todomate-spa",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "vite": "^4.3.2"
+  }
+}
