+++ ToDoMate-SPA/src/App.js
@@ -1,0 +1,8 @@
+import Router from "./components/router";
+
+function App($container) {
+  this.$container = $container;
+
+  new Router($container);
+}
+export default App;
