+++ ToDoMate-SPA/src/index.js
@@ -1,0 +1,5 @@
+import App from "./App";
+
+window.addEventListener("DOMContentLoaded", () => {
+  new App(document.querySelector("#app"));
+});
