+++ ToDoMate-SPA/src/styles/style.css
@@ -19,6 +19,7 @@ body {
 /* header */
 .header {
   display: flex;
+  justify-content: center;
   align-items: center;
 
   height: 6rem;
