+++ ToDoMate-SPA/todomate.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="ko">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="style.css" />
+    <link rel="stylesheet" href="../reset.css" />
+    <script defer type="module" src="./TODO_DATA.js"></script>
+    <script defer type="module" src="./todolist.js"></script>
+    <title>To Do Mate</title>
+  </head>
+  <body>
+    <div id="app"></div>
+
+    <footer class="footer">
+      <a href="/">달력</a>
+      <a href="/mycategory">MY</a>
+    </footer>
+  </body>
+</html>
