--- ToDoMate/mycategory.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="../reset.css" />
-    <link rel="stylesheet" href="mycategory.css" />
-    <script defer type="module" src="./TODO_DATA.js"></script>
-    <script defer type="module" src="./draganddrop.js"></script>
-    <title>My Page</title>
-  </head>
-  <body>
-    <header class="header">
-      <h1>WEB TO DO MATE</h1>
-    </header>
-    <main>
-      <section class="mycategory__section">
-        <h2>나의 카테고리</h2>
-        <ul></ul>
-      </section>
-    </main>
-    <footer class="footer">
-      <a href="todomate.html">달력</a>
-      <a href="mycategory.html">MY</a>
-    </footer>
-  </body>
-</html>
